+++ R/SimWarnings.R
@@ -8,6 +8,9 @@
 #' @param subset logical; take a subset of the \code{design} object showing only conditions that
 #'   returned warnings?
 #'
+#' @param fuzzy logical; group warning messages that differ only slightly into a single
+#'   column? Set to \code{FALSE} to report every distinct message separately
+#'
 #' @references
 #'
 #' Chalmers, R. P., & Adkins, M. C.  (2020). Writing Effective and Reliable Monte Carlo Simulations
@@ -75,11 +78,11 @@
 #' SimWarnings(res, subset=FALSE)
 #'
 #'
-SimWarnings <- function(obj, subset=TRUE){
+SimWarnings <- function(obj, subset=TRUE, fuzzy=TRUE){
     if(!any(colnames(obj) == 'WARNINGS')) return(dplyr::tibble())
     warnings <- obj$WARNINGS
     pick <- which(warnings > 0)
-    ret <- SimExtract(obj, what='warnings')
+    ret <- SimExtract(obj, what='warnings', fuzzy=fuzzy)
     if(subset) ret <- ret[pick, ]
     ret
 }